Move concept models to ontology package
