+++ s3/client.go
@@ -2,12 +2,11 @@ package s3
 
 import (
 	"context"
+	"encoding/json"
 	"net"
 	"net/http"
 	"strings"
 	"time"
-
-	"encoding/json"
 
 	fthealth "github.com/Financial-Times/go-fthealth/v1_1"
 	"github.com/Financial-Times/go-logger"
@@ -15,10 +14,12 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3"
+
+	"github.com/Financial-Times/aggregate-concept-transformer/ontology"
 )
 
 type Client interface {
-	GetConceptAndTransactionID(ctx context.Context, UUID string) (bool, Concept, string, error)
+	GetConceptAndTransactionID(ctx context.Context, UUID string) (bool, ontology.Concept, string, error)
 	Healthcheck() fthealth.Check
 }
 
@@ -60,7 +61,7 @@ func NewClient(bucketName string, awsRegion string) (Client, error) {
 	}, err
 }
 
-func (c *ConceptClient) GetConceptAndTransactionID(ctx context.Context, UUID string) (bool, Concept, string, error) {
+func (c *ConceptClient) GetConceptAndTransactionID(ctx context.Context, UUID string) (bool, ontology.Concept, string, error) {
 	getObjectParams := &s3.GetObjectInput{
 		Bucket: aws.String(c.bucketName),
 		Key:    aws.String(getKey(UUID)),
@@ -71,10 +72,10 @@ func (c *ConceptClient) GetConceptAndTransactionID(ctx context.Context, UUID str
 		e, ok := err.(awserr.Error)
 		if ok && e.Code() == "NoSuchKey" {
 			// NotFound rather than error, so no logging needed.
-			return false, Concept{}, "", nil
+			return false, ontology.Concept{}, "", nil
 		}
 		logger.WithError(err).WithUUID(UUID).Error("Error retrieving concept from S3")
-		return false, Concept{}, "", err
+		return false, ontology.Concept{}, "", err
 	}
 
 	getHeadersParams := &s3.HeadObjectInput{
@@ -84,14 +85,14 @@ func (c *ConceptClient) GetConceptAndTransactionID(ctx context.Context, UUID str
 	ho, err := c.s3.HeadObjectWithContext(ctx, getHeadersParams)
 	if err != nil {
 		logger.WithError(err).WithUUID(UUID).Error("Cannot access S3 head object")
-		return false, Concept{}, "", err
+		return false, ontology.Concept{}, "", err
 	}
 	tid := ho.Metadata["Transaction_id"]
 
-	var concept Concept
+	var concept ontology.Concept
 	if err = json.NewDecoder(resp.Body).Decode(&concept); err != nil {
 		logger.WithError(err).WithUUID(UUID).Error("Cannot unmarshal object into a concept")
-		return true, Concept{}, "", err
+		return true, ontology.Concept{}, "", err
 	}
 	return true, concept, *tid, nil
 }
